# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SME Finance Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{profileId}/position:
    get:
      tags: [Position]
      operationId: getFinancialPosition
      summary: Get the computed financial position for a profile
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
      responses:
        '200':
          description: Computed financial position.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialPosition'
...
  /profiles/{profileId}/business-ideas:
    get:
      tags: [Position]
      operationId: listBusinessIdeas
      summary: List business ideas with live financial baselines
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
      responses:
        '200':
          description: List of business ideas with live baselines.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessIdea'
...
components:
  parameters:
...
    ProfileIdParam:
      name: profileId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    PLBreakdown:
      type: object
      required: [revenues, confirmedExpenses, pendingExpenses, profit]
      properties:
        revenues:
          type: number
        confirmedExpenses:
          type: number
        pendingExpenses:
          type: number
        profit:
          type: number
    TaxLine:
      type: object
      required: [label, amount]
      properties:
        label:
          type: string
        amount:
          type: number
    TaxCalculation:
      type: object
      required: [lines, balanceDue, reserveGap]
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/TaxLine'
        balanceDue:
          type: number
        reserveGap:
          type: number
    AccountBalance:
      type: object
      required: [name, balance]
      properties:
        name:
          type: string
        balance:
          type: number
    CashPosition:
      type: object
      required: [accounts, taxReserve, apDueWithin30Days, netAvailable]
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountBalance'
        taxReserve:
          type: number
        apDueWithin30Days:
          type: number
        netAvailable:
          type: number
    AREntry:
      type: object
      required: [name, amount]
      properties:
        name:
          type: string
        amount:
          type: number
        daysPastDue:
          type: number
        invoiceCount:
          type: number
    APEntry:
      type: object
      required: [name, amount]
      properties:
        name:
          type: string
        amount:
          type: number
        daysUntilDue:
          type: number
    KPI:
      type: object
      required: [id, label, value, trend, basis]
      properties:
        id:
          type: string
        label:
          type: string
        value:
          type: string
        trend:
          type: string
        basis:
          type: string
        rawValue:
          type: number
        detail:
          type: string
    SAReadiness:
      type: object
      required: [score, completedCount, totalCount]
      properties:
        score:
          type: number
        completedCount:
          type: number
        totalCount:
          type: number
    FinancialPosition:
      type: object
      required: [plBreakdown, taxCalculation, cashPosition, arEntries, apEntries, kpis, saReadiness, pendingInboxCount]
      properties:
        plBreakdown:
          $ref: '#/components/schemas/PLBreakdown'
        taxCalculation:
          $ref: '#/components/schemas/TaxCalculation'
        cashPosition:
          $ref: '#/components/schemas/CashPosition'
        arEntries:
          type: array
          items:
            $ref: '#/components/schemas/AREntry'
        apEntries:
          type: array
          items:
            $ref: '#/components/schemas/APEntry'
        kpis:
          type: array
          items:
            $ref: '#/components/schemas/KPI'
        saReadiness:
          $ref: '#/components/schemas/SAReadiness'
        pendingInboxCount:
          type: number
...
    ImpactRange:
      type: object
      required: [min, max]
      properties:
        min:
          type: number
        max:
          type: number
    PaybackRange:
      type: object
      properties:
        minMonths:
          type: [number, 'null']
        maxMonths:
          type: [number, 'null']
    AssumptionField:
      type: object
      required: [key, label, value, unit, min, max, step]
      properties:
        key:
          type: string
        label:
          type: string
        value:
          type: number
        unit:
          type: string
        min:
          type: number
        max:
          type: number
        step:
          type: number
    BusinessIdea:
      type: object
      required: [id, category, title, summary, currentPosition, proposedAction, priorityTier, source, confidence, status, editableAssumptions, whatMustBeTrue]
      properties:
        id:
          type: string
        category:
          type: string
        title:
          type: string
        summary:
          type: string
        currentPosition:
          type: string
        proposedAction:
          type: string
        priorityTier:
          type: string
        plImpactRange:
          $ref: '#/components/schemas/ImpactRange'
        cashImpactRange:
          $ref: '#/components/schemas/ImpactRange'
        taxImpactRange:
          $ref: '#/components/schemas/ImpactRange'
        paybackRange:
          $ref: '#/components/schemas/PaybackRange'
        urgencyNote:
          type: string
        editableAssumptions:
          type: array
          items:
            $ref: '#/components/schemas/AssumptionField'
        whatMustBeTrue:
          type: array
          items:
            type: string
        source:
          type: string
        confidence:
          type: string
        status:
          type: string
        committedDecisionId:
          type: [string, 'null']
...